requirements.txt
tqdm
pandas>=2.1.2, <2.2.0
gluonts>=0.16.0
autogluon.timeseries>=1.2
tabpfn-client>=0.1.7

# tabpfn>=2.0.9
-e ./tabpfn_time_series/experimental/finetuning/tabpfn_private

datasets>=3.3.2
python-dotenv>=1.1.0

# for python 3.12
pyyaml>=6.0.1
